- Ajout de commentaires - Bugfix du fait que l'on pouvait marcher sur une sortie puis prendre la donnée et gagner instantannément  -Ajout de l'entrechoc entre les drones (deux drones de même type se désactivent lorsqu'ils s'entrechoquent

+++ src/gui/ImgParcelle.java
@@ -53,7 +53,7 @@ public class ImgParcelle  {
 
     public Parcelle getParcelle() { return parcelle; }
 
-    public void setOpacity(int i) {
+    public void setOpacity(double i) {
         elt.setOpacity(i);
     }
 }
